Update README.md
This project is a web-based application that predicts the quality of red wine as either "Good" or "Bad" based on various input features. It utilizes a Python-based machine learning model to process user input values (e.g., acidity, alcohol content, pH, etc.) and deliver a prediction on the wine's quality. The website is built with HTML/CSS for the front end and integrates Python for the prediction model using flask.